third_party/blink/web_tests/fast/js/string-prototype-properties-expected.txt

This is a test case for bug 64677.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS String.prototype.toString.call(undefined) threw exception TypeError: String.prototype.toString requires that 'this' be a String.
PASS String.prototype.valueOf.call(undefined) threw exception TypeError: String.prototype.valueOf requires that 'this' be a String.
PASS String.prototype.charAt.call(undefined, 0) threw exception TypeError: String.prototype.charAt called on null or undefined.
PASS String.prototype.charCodeAt.call(undefined, 0) threw exception TypeError: String.prototype.charCodeAt called on null or undefined.
PASS String.prototype.concat.call(undefined, 'five') threw exception TypeError: String.prototype.concat called on null or undefined.
PASS String.prototype.indexOf.call(undefined, '2') threw exception TypeError: String.prototype.indexOf called on null or undefined.
PASS String.prototype.lastIndexOf.call(undefined, '2') threw exception TypeError: String.prototype.lastIndexOf called on null or undefined.
PASS String.prototype.match.call(undefined, /2+/) threw exception TypeError: String.prototype.match called on null or undefined.
PASS String.prototype.replace.call(undefined, /2+/, '-') threw exception TypeError: String.prototype.replace called on null or undefined.
PASS String.prototype.search.call(undefined, '4') threw exception TypeError: String.prototype.search called on null or undefined.
PASS String.prototype.slice.call(undefined, 1, 3) threw exception TypeError: String.prototype.slice called on null or undefined.
PASS String.prototype.split.call(undefined, '2') threw exception TypeError: String.prototype.split called on null or undefined.
PASS String.prototype.slice.call(undefined, 1, 3) threw exception TypeError: String.prototype.slice called on null or undefined.
PASS String.prototype.substr.call(undefined, 1, 3) threw exception TypeError: String.prototype.substr called on null or undefined.
PASS String.prototype.substring.call(undefined, 1, 3) threw exception TypeError: String.prototype.substring called on null or undefined.
PASS String.prototype.toLowerCase.call(undefined) threw exception TypeError: String.prototype.toLowerCase called on null or undefined.
PASS String.prototype.toUpperCase.call(undefined) threw exception TypeError: String.prototype.toUpperCase called on null or undefined.
PASS String.prototype.localeCompare.call(undefined, '1224') threw exception TypeError: String.prototype.localeCompare called on null or undefined.
PASS String.prototype.toLocaleLowerCase.call(undefined) threw exception TypeError: String.prototype.toLocaleLowerCase called on null or undefined.
PASS String.prototype.toLocaleUpperCase.call(undefined) threw exception TypeError: String.prototype.toLocaleUpperCase called on null or undefined.
PASS String.prototype.trim.call(undefined) threw exception TypeError: String.prototype.trim called on null or undefined.
PASS String.prototype.toString.call(1224) threw exception TypeError: String.prototype.toString requires that 'this' be a String.
PASS String.prototype.valueOf.call(1224) threw exception TypeError: String.prototype.valueOf requires that 'this' be a String.
PASS String.prototype.charAt.call(1224, 0) is "1"
PASS String.prototype.charCodeAt.call(1224, 0) is 0x31
PASS String.prototype.concat.call(1224, 'five') is "1224five"
PASS String.prototype.indexOf.call(1224, '2') is 1
PASS String.prototype.lastIndexOf.call(1224, '2') is 2
PASS String.prototype.match.call(1224, /2+/) is ["22"]
PASS String.prototype.replace.call(1224, /2+/, '-') is "1-4"
PASS String.prototype.search.call(1224, '4') is 3
PASS String.prototype.slice.call(1224, 1, 3) is "22"
PASS String.prototype.split.call(1224, '2') is ["1","","4"]
PASS String.prototype.slice.call(1224, 1, 3) is "22"
PASS String.prototype.substr.call(1224, 1, 3) is "224"
PASS String.prototype.substring.call(1224, 1, 3) is "22"
PASS String.prototype.toLowerCase.call(1224) is "1224"
PASS String.prototype.toUpperCase.call(1224) is "1224"
PASS String.prototype.localeCompare.call(1224, '1224') is 0
PASS String.prototype.toLocaleLowerCase.call(1224) is "1224"
PASS String.prototype.toLocaleUpperCase.call(1224) is "1224"
PASS String.prototype.trim.call(1224) is "1224"
PASS successfullyParsed is true

TEST COMPLETE
